+++ pajbot/bot.py
@@ -655,6 +655,10 @@ class Bot:
 
     def ban_login(self, login: str, reason: Optional[str] = None) -> None:
         user_id = self.twitch_helix_api.get_user_id(login)
+        if user_id is None:
+            log.error(f"Attempted to ban user with login {login}, but no such user was found")
+            return
+
         if self._has_moderation_actions():
             self.thread_locals.moderation_actions.add(login, Ban(reason))
         else:
@@ -687,14 +691,50 @@ class Bot:
         else:
             self.unban_id(user_id)
 
+    def _untimeout(self, user_id: str) -> None:
+        try:
+            ban_data = self.twitch_helix_api.get_banned_user(
+                self.streamer.id, self.streamer_access_token_manager, user_id
+            )
+            if ban_data is None:
+                log.warning(f"User with ID {user_id} is already not banned or timed-out.")
+                return
+            # As per the Helix docs, expires_at will return empty if the user is permabanned.
+            if not ban_data.expires_at:
+                log.error(f"User with ID {user_id} is currently banned! Will not untimeout.")
+                return
+        except HTTPError as e:
+            if e.response.status_code == 401:
+                log.error(f"Failed to get banned user status with id {user_id}, unauthorized: {e} - {e.response.text}")
+                return
+            else:
+                log.error(f"Failed to get banned user status with id {user_id}: {e} - {e.response.text}")
+                return
+
+        try:
+            self.twitch_helix_api.unban_user(self.streamer.id, self.bot_user.id, user_id, self.bot_token_manager)
+        except HTTPError as e:
+            if e.response.status_code == 401:
+                log.error(f"Failed to untimeout user with id {user_id}, unauthorized: {e} - {e.response.text}")
+            else:
+                log.error(f"Failed to untimeout user with id {user_id}: {e} - {e.response.text}")
+
     def untimeout(self, user: User) -> None:
-        self.untimeout_login(user.login)
+        self.untimeout_id(user.id)
+
+    def untimeout_id(self, user_id: str) -> None:
+        self._untimeout(user_id)
 
     def untimeout_login(self, login: str) -> None:
+        user_id = self.twitch_helix_api.get_user_id(login)
+        if user_id is None:
+            log.error(f"Attempted to untimeout user with login {login}, but no such user was found")
+            return
+
         if self._has_moderation_actions():
             self.thread_locals.moderation_actions.add(login, Untimeout())
         else:
-            self.privmsg(f"/untimeout {login}")
+            self.untimeout_id(user_id)
 
     def _timeout(self, login: str, duration: int, reason: Optional[str] = None) -> None:
         message = f"/timeout {login} {duration}"